Add the final improvements before tagging version 0.1.0-alpha0.
* Features
Change the "source files" to the "input files."
Cover C++ and assembly as well. (Configuration only, building not tested.)
Support the customization on the input file extentions.
Support customizing the filename extension of the Rust FFI binding files.

* Licensing
Add the contributors to the copyright holders.
Clarify the licensing of the items imported via the Rust FFI.

* Documents
Brush up the documents.